+++ src/lib/lambda/index.ts
@@ -1,7 +1,7 @@
 
 // ---
 // lol how do i even do named exports
-import { parseTerm } from './parser2';
+import { parseTerm } from './parser';
 import { renderExpression } from './renderer';
 import { renderAsChurchNumeral, renderAsChurchBoolean } from './churchPrimitives';
 import { getFreeVars } from './util';
