+++ src/lib/lambda/renderer.ts
@@ -9,12 +9,12 @@ function renderExpression(expression) {
       }
       // I have no idea whether the rendering of the right side is valid.
       let rightSide;
-      //if(expression.right.type !== 'application'){
+      if(expression.right.type !== 'application'){
         rightSide = renderExpression(expression.right)
-      //} else {
-      //  rightSide = `(${renderExpression(expression.right)})`
-      //}
-      return `(${leftSide}${rightSide})`;
+      } else {
+        rightSide = `(${renderExpression(expression.right)})`
+      }
+      return `${leftSide}${rightSide}`;
     case 'function':
       return `λ${expression.argument}.${renderExpression(expression.body)}`
     case 'token':
